+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this package adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.2-pre.2] - 2022-01-13
+### Fixed
+- Support for HTML character codes in Rich Text Parser (&amp; etc.)
+- Fixed how UnityEditor.UIElementsModule is fetched in UIElementsHelper for latest editor versions.
+
 ## [2.1.1] - 2021-11-02
 ### Changed
 - Minor documentation update.